MB-59625: fix availableHosts calculation
context:
UpdateConnPools() subscribe to the event FtsTopologyChange.
Upon which, it
> calculates addedHosts and removedHosts
> Retry reaching out to previously unavailable hosts.
  ( Producing nowAvailableHosts)
> Create conn pools for addedHosts and nowAvailableHosts
> Append them to list availableHosts
> Close pools corresponding to removedHosts

This issue is caused by a case where a host is commonly present
in removedHosts and nowAvailableHosts list.
Because of which, finally the pool was closed for the host, but
it was added to availableHosts list.
As a result, some queries were served by this host, and they
failed as conn pool doesn't exist.

This change:
 * Wrap the method responsible for fetching host credential
   into a exponential backoff retry loop.
 * Avoid adding a host to list of availableHosts, if it is present
   in both removedHosts and nowAvailableHosts.

Change-Id: I2dbf4cde10d667c4a48cc69e6392e1ff45a63dc8
Reviewed-on: https://review.couchbase.org/c/n1fty/+/201096
Tested-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>