MB-59625 fix cfg subscription
One of the major event for updateConnPools
is topologyChange notification. In the absence
of an indexer, pools won't be materialized. But,
we still need to track topologyChange notifications
for book keeping of list of hosts and availableHosts.

FTS Index creation/deletion events are also tied up with
conn pools, as they trigger regiter/unregister indexer call.

Thus, we need to subscribe to cfg metaKV events
early on (at init time).

Change-Id: I2c0fa565c97c3208f6d8fb9d8d752f9d1f9d1cb2
Reviewed-on: https://review.couchbase.org/c/n1fty/+/201884
Tested-by: Mohd Shaad Khan <mohd.shaadkhan@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>