MB-65809 Add Infrastructure to check and create query meta data bucket
Add infrastructure similar to existing mechansim to check and
create N1QL_SYSTEM_BUCKET when necessary, for the new query
meta data bucket. The previous infrastructure is tied to the
dictionary (keeping optimizer statistics), and the new query
meta data bucket is added there too, since the code is very
similar. The plan stability feature which currently uses the
new query meta data bucket is not related to the dictionary,
but since the dictionary will be automatically initialized
(in enterprise edition) anyway, the query meta data bucket
is piggy backed there as well.

Change-Id: Ibc019a7d3c94a818b45021cdc50e8940b104828d
Reviewed-on: https://review.couchbase.org/c/n1fty/+/236255
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>